refactor: pass NetworkSample to the bound and keep loss as f64
bound::Reading repeated NetworkSample minus the round trip fields, and
loss went from f64 to f32 in the facade and back to f64 for the bound.
The bound now reads the sample itself, and NetworkSample::loss is f64
like the figure the link monitor measures.

+++ iroh-live-egui/src/overlay.rs
@@ -591,7 +591,7 @@ fn net_playback(network: Option<&NetworkSample>, link: &[String]) -> Section {
         lines.push(Line::info(format!("min rtt: {:.1} ms", millis(min_rtt))));
     }
     if let Some(loss) = net.loss {
-        let pct = f64::from(loss) * 100.0;
+        let pct = loss * 100.0;
         parts.push(format!("loss:{pct:.1}%"));
         let color = lower_is_better(pct, 2.0, 10.0);
         lines.push(Line::metric(
